Crate objc_test_utils

Enums

AddBlock
A block that takes one integer argument, adds to it, and returns the sum: int32_t (^)(int32_t).

IntBlock
A block that takes no arguments and returns an integer: int32_t (^)().

Functions

get_add_block
Returns a pointer to a global AddBlock that returns its argument + 7.

get_add_block_with
Returns a pointer to a copied AddBlock that returns its argument + i.

get_int_block
Returns a pointer to a global IntBlock that returns 7.

get_int_block_with
Returns a pointer to a copied IntBlock that returns i.

invoke_add_block
Invokes an AddBlock with a and returns the result.

invoke_int_block
Invokes an IntBlock and returns its result.